Committing in .
 changed one parameter in the definition of wxBitmap from long to int.
 (just as it is in wxGTK). This solves a compilation problem on OpenVMS.

 Modified Files:
 	wxWindows/include/wx/motif/bitmap.h
 	wxWindows/src/motif/bitmap.cpp
 ----------------------------------------------------------------------


git-svn-id: https://svn.wxwidgets.org/svn/wx/wxWidgets/trunk@16275 c3d73ce0-8a6f-49c7-b76d-6d57e0e08775

+++ include/wx/motif/bitmap.h
@@ -149,7 +149,7 @@ public:
     wxBitmap(char **data, wxControl* control);
     
     // Load a file or resource
-    wxBitmap(const wxString& name, long type = wxBITMAP_TYPE_XPM);
+    wxBitmap(const wxString& name, int type = wxBITMAP_TYPE_XPM);
     
     // Constructor for generalised creation from data
     wxBitmap(void *data, long type, int width, int height, int depth = 1);
